Renamed java-package containing JavaScript handler
