+++ javascript/src/main/java/dk/dbc/search/work/presentation/javascript/JavascriptCacheObjectBuilder.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package dk.dbc.search.work.presentation;
+package dk.dbc.search.work.presentation.javascript;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
