+++ worker/src/main/java/dk/dbc/search/work/presentation/worker/ParallelCacheContentBuilder.java
@@ -18,7 +18,7 @@
  */
 package dk.dbc.search.work.presentation.worker;
 
-import dk.dbc.search.work.presentation.JavascriptCacheObjectBuilder;
+import dk.dbc.search.work.presentation.javascript.JavascriptCacheObjectBuilder;
 import dk.dbc.search.work.presentation.api.jpa.CacheEntity;
 import dk.dbc.search.work.presentation.api.jpa.WorkContainsEntity;
 import dk.dbc.search.work.presentation.api.pojo.ManifestationInformation;
