+++ worker/src/main/java/dk/dbc/search/work/presentation/worker/tree/CacheContentBuilder.java
@@ -18,7 +18,7 @@
  */
 package dk.dbc.search.work.presentation.worker.tree;
 
-import dk.dbc.search.work.presentation.JavascriptCacheObjectBuilder;
+import dk.dbc.search.work.presentation.javascript.JavascriptCacheObjectBuilder;
 import dk.dbc.search.work.presentation.api.pojo.ManifestationInformation;
 import dk.dbc.search.work.presentation.worker.CorepoContentServiceConnector;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
